Make CI work (#97)
Same PR as #95 but post repo move

This uses GitHub Actions and haskell-ci. I decided not to figure out
CircleCI.

You do need the latest haskell-ci version from git to be able to
generate for GHC 9.4.2. But it works!

This also builds and tests with GHC 9.4.2 ;-)

It also comes with a bonus Nix flake, because I can't help myself I
guess 😆

Fixes #94

Co-authored-by: Alexander Vieth <aovieth@gmail.com>