Do paged conversation lists; change the pager to be generic over types (#105)
I need to be able to do paged conversation lists for slacklinker, since
I expect I will have a lot of conversations to go through.

This meant that the existing pager stuff that was hardcoded to one
response type needed to fixed. So I fixed it by deleting a bunch of
duplicated code by making it generic.