Fix shared-channel Conversation parse; improve Conversation parse errors
This bug was found using the conversationsList endpoint, in which our parse
failed on a shared channel.

I have committed an anonymized version of the bad part of the response
as a regression test, and fixed the error messages for Conversations
matching a variant but failing to parse that variant.