Support for Block Kit header and button accessory (#116)
## Context

In order to create a more complex slack message, I needed to add support
for the Block Kit Header element, and add optional Accessories to
Section elements. Headers are a short but high-level, large, bold
plaintext element. Accessories allow Block Elements to float to the
right of a Section block, to provide small buttons and images close to
the section they are related to. Accessories provide a better usage of
horizontal space in messages, and prevent using another vertical block
row for an element where unnecessary.

## Documentation

### Relevant Slack API Links
[Header Blocks](https://api.slack.com/reference/block-kit/blocks#header)
[Section
Blocks](https://api.slack.com/reference/block-kit/blocks#section)
Field | Type | Required? | Description
-- | -- | -- | --
accessory | Object | No | One of the available [element
objects](https://api.slack.com/reference/messaging/block-elements).

[List of Block Elements, also called Element Objects, e.g.
buttons](https://api.slack.com/reference/block-kit/block-elements)
> Block elements can be used inside of
[section](https://api.slack.com/reference/block-kit/blocks#section),
[context](https://api.slack.com/reference/block-kit/blocks#context),
[input](https://api.slack.com/reference/block-kit/blocks#input) and
[actions](https://api.slack.com/reference/block-kit/blocks#actions)
[layout blocks](https://api.slack.com/reference/block-kit/blocks).

### Code Feature Changes
- Header blocks are implemented as a core `SlackBlock` type, accepting
one `SlackPlainTextOnly` element, with a comment noting a `maxlength` of
150, which is done similarly to other implemented types.
- Section blocks now also take a `Maybe SlackAccessory`, to allow the
option of tagging one on to the right of a section of text.
- Slack accessories were added as a new datatype, which can only be a
`SlackButtonAccessory`, which takes a `SlackAction` (button).
- **Currently, accessories only provide support for buttons.** Buttons
are defined consistently by the API for use in both action lists (as
already implemented,) and in the context of sections as an accessory, so
no changes were necessary to them themselves.
- Support for using images as accessories to sections should be simple
enough to implement as well, given the existing implementation of
`SlackBlockImage` which provides the fully-implemented type
`SlackImage`, but was not within my scope of work at this time.
- For the scope of my work, Slack is not currently desired to write any
data out from messages for security/compliance concerns. Since the
majority of block elements are interactive in some way that would write
_from_ Slack _through_ the API, support for them is not provided at this
time.


## Effects

- **Breaking Change!** Any uses of the type `SlackBlockSection` provided
by this repo will now need to update their code to reflect the addition
of the `Maybe SlackAccessory` which was not previously defined.
Providing a simple `Nothing` after the `SlackText` provided to the data
constructor is sufficient. Given that this code is marked Experimental
and development is ongoing, changes like this should not be too
surprising.
- ~_This repo does not provide coded nor documentation of tests for
Block Kit elements._~

Co-authored-by: Matthew Mongeau <halogenandtoast@gmail.com>
Co-authored-by: Jade Lovelace <jadel@mercury.com>