DUX-3346: Support custom deserializers in request verification
This turns out to be necessary for the impressively messed up nonsense
that Slack does with their interactive webhooks, which range from "json
in a form parameter" (WHAT) to "entirely putting the thing in form
parameters".

More code to support this may come in the future, but we need this
foundation to be able to disentangle the bad code in mwb using this.